workflow/bgcflow/requirements.txt
# local package
-e .

# external requirements
biopython
pandas
subprocess
datetime
pathlib
alive_progress